AI/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
cssselect2==0.9.0
distro==1.9.0
fastapi==0.129.0
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
kombu==5.6.2
lxml==6.0.2
MarkupSafe==3.0.3
openai==2.21.0
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
prompt_toolkit==3.0.52
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pydyf==0.12.1
pypdf==6.7.0
pypdfium2==5.4.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
redis==7.2.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
